Saw that different seeds can give very different probabilities.  This commit is just a reminder that how the pseudo-random number generator is initialized can affect the outcome by quite a bit.
In the MDPBidderUAI class, use the empirical transition probabilities instead of distributions over highest other prices.  Changed the way the histogram is computed in the discrete case to use all of the learned prices, since they're being discretized.  In the transition matrix function, check if we need to perform interpolation first.  The value that needs to be calculated might already be stored.